requirements.txt
Babel==2.11.0
beautifulsoup4==4.11.1
bs4==0.0.1
click==8.1.3
docx==0.2.4
docx-utils==0.1.3
docx2pdf==0.1.8
docxcompose==1.3.7
docxtpl==0.16.4
duplicates==0.1.0
iteration-utilities==0.11.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.5
Pillow==9.3.0
PyPDF2==2.11.2
python-docx==0.8.11
python-utils==3.4.5
pytz==2022.6
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.1
xmltodict==0.13.0
asgiref==3.5.2
Django==4.1.3
image==1.5.33
opencv-contrib-python==4.6.0.66
opencv-contrib-python-headless==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
pytesseract==0.3.10
scipy==1.9.3
sqlparse==0.4.3
alembic==1.8.1 
distlib==0.3.6
filelock==3.8.2
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
python-dateutil==2.8.2
python-decouple==3.6
psycopg2-binary==2.9.5
setuptools==59.6.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
wheel==0.37.1
